Create Handler class
